Allow setting of the generator string with an env var
In the case of running a cmake job on multiple nodes with a matrix project, it
is useful to have a single jenkins job trigger builds with different generators
depending on the resources available on the node (Visual Studio versions, xcode,
makefiles etc). With this patch, each node can specify a generator to be used
when building cmake jobs on it (or anywhere else an environment var can be
defined).